+++ packages/common/src/components/index.ts
@@ -1,4 +1,5 @@
 import { CommonContextProvider as CommonContextProviderComponent } from './common_context';
+import { ErrorFallback as ErrorFallbackComponent } from './error_handler';
 import { MDXRenderer as MDXRendererComponent } from "./mdx";
 import { PythonKorea as PythonKoreaComponent } from './pythonkorea';
 
@@ -6,6 +7,7 @@ namespace Components {
   export const CommonContextProvider = CommonContextProviderComponent;
   export const MDXRenderer = MDXRendererComponent;
   export const PythonKorea = PythonKoreaComponent;
+  export const ErrorFallback = ErrorFallbackComponent;
 }
 
 export default Components;
